Proto: Circumvent a bug in OCaml 4.13 related to tailcall annotations
As reported in https://github.com/ocaml/ocaml/issues/10507 and as long
as https://github.com/ocaml/ocaml/pull/10508 is not merged, the
operators `( @@ )` and `( |> )` do not play well with the
`ocaml.tailcall` annotation. Unfortunately, the Michelson operator
makes a heavy use of both.

Despite the absolute superiority of `@` over parentheses in these
circumstances, we need to alter the aesthetic of the original code to
prepare the use of OCaml 4.13 with Tezos.