Proto,tx_rollup: reinstate test on burn_per_byte
As stated in !4200, it is better to use the built-in mechanism of
`conv_with_guard` rather than to use `assert` when checking properties
on values during a json decoding.

This commit turns the latter into the former, in the encoding of
`proto_alpha.lib_protocol.tx_rollup_state_repr.t`

NB:
The test has been disabled without motivation in commit
`Proto,tx_rollup: Store a linked list of inbox levels` (e93cb0ee), as
part of !4332. This commit reinstates that test.

Closes #2358