Tezt,tx_rollup: Fix flakiness of [check_injection]
When the injector flushes its batches of layer-1 operations to inject,
it emits one “injecting” event per layer-1 operation, then one
“completed” event.

[check_injection] is a function that checks that an “injecting” event
is emitted before a “completed” event.

The previous implementation of [check_injection] was flaky, because it
was making an implicit hypothesis that the “completed” event was
emitted between two calls of [check_injection]. As it turns out, it is
not always the case.

To fix this flakiness, we do two things:

   1. We only wait for the “injecting” event, but we set up a timeout
     (previously, the code was waiting for both the “injecting” and
     “completed” events, and if the latter came before the former, it
     meant the “injecting” event was never throw.
   2. After “consuming” an “injecting” event, we also consume the
      “completed” event explicitly.